+++ odesk/holidaysrating/application/views/admin/pictures.php
@@ -4,9 +4,15 @@
     <?php foreach($pictures as $picture): ?>
     <?php $image = array('src' => './uploads/media/'.$picture->name, 'class' => 'thumbnail') ?>
     <li>
-        <a href="<?php echo base_url() ?>uploads/media/<?php echo $picture->name ?>"><?php echo img($image) ?></a>
-        <a href="<?php echo base_url() ?>index.php/admin/picture_delete/<?php echo $picture->id ?>"><i class="icon-remove"></i> Delete</a>
+        <a class="fancybox" href="<?php echo base_url() ?>uploads/media/<?php echo $picture->name ?>"><?php echo img($image) ?></a>
+        <a href="<?php echo base_url() ?>index.php/admin/picture_delete/<?php echo $picture->id ?>" onclick="return confirm('You are about to delete a record. This cannot be undone. Are you sure?')"><i class="icon-remove"></i> Delete</a>
     </li>
     <?php endforeach ?>
 </ul>
 <?php echo anchor('admin/index', 'Back') ?>
+
+<script type="text/javascript">
+	$(document).ready(function() {
+		$(".fancybox").fancybox();
+	});
+</script>
